# Optimize ReplaceLineEndings