@MihaZupan 2023-02-05
Comparing the combined change of #81630 + #80963 + #78678 against main without them (NET 7 vs 8):

ReplaceLineEndings LF => LF

| Toolchain | NewLineFrequency | Length | Source | Replacement |         Mean | Ratio | Allocated |
|---------- |----------------- |------- |------- |------------ |-------------:|------:|----------:|
|      main |             0.05 |  10000 |     LF |          LF |  28,130.2 ns |  1.00 |   20024 B |
|        pr |             0.05 |  10000 |     LF |          LF |     623.1 ns |  0.02 |         - |
|           |                  |        |        |             |              |       |           |
|      main |              0.1 |  10000 |     LF |          LF |  45,561.1 ns |  1.00 |   20024 B |
|        pr |              0.1 |  10000 |     LF |          LF |     607.1 ns |  0.01 |         - |
|           |                  |        |        |             |              |       |           |
|      main |              0.2 |  10000 |     LF |          LF |  79,684.7 ns | 1.000 |   20024 B |
|        pr |              0.2 |  10000 |     LF |          LF |     607.1 ns | 0.008 |         - |
|           |                  |        |        |             |              |       |           |
|      main |                1 |  10000 |     LF |          LF | 252,398.5 ns | 1.000 |   20024 B |
|        pr |                1 |  10000 |     LF |          LF |     620.9 ns | 0.002 |         - |

<details>
<summary>ReplaceLineEndings CRLF => LF</summary>

| Toolchain | NewLineFrequency | Length | Source | Replacement |         Mean | Ratio | Allocated |
|---------- |----------------- |------- |------- |------------ |-------------:|------:|----------:|
|      main |             0.05 |  10000 |   CRLF |          LF |  26,852.0 ns |  1.00 |   19040 B |
|        pr |             0.05 |  10000 |   CRLF |          LF |   8,702.7 ns |  0.32 |   19040 B |
|           |                  |        |        |             |              |       |           |
|      main |              0.1 |  10000 |   CRLF |          LF |  41,837.3 ns |  1.00 |   18128 B |
|        pr |              0.1 |  10000 |   CRLF |          LF |  16,860.2 ns |  0.40 |   18128 B |
|           |                  |        |        |             |              |       |           |
|      main |              0.2 |  10000 |   CRLF |          LF |  67,020.5 ns |  1.00 |   16664 B |
|        pr |              0.2 |  10000 |   CRLF |          LF |  29,707.1 ns |  0.44 |   16664 B |
|           |                  |        |        |             |              |       |           |
|      main |                1 |  10000 |   CRLF |          LF | 136,184.9 ns |  1.00 |   10024 B |
|        pr |                1 |  10000 |   CRLF |          LF |  58,194.7 ns |  0.43 |   10024 B |

</details>

<details>
<summary>ReplaceLineEndings CRLF => CRLF</summary>

| Toolchain | NewLineFrequency | Length | Source | Replacement |       Mean | Ratio | Allocated |
|---------- |----------------- |------- |------- |------------ |-----------:|------:|----------:|
|      main |             0.05 |  10000 |   CRLF |        CRLF |  28.686 us |  1.00 |   20024 B |
|        pr |             0.05 |  10000 |   CRLF |        CRLF |   6.715 us |  0.23 |         - |
|           |                  |        |        |             |            |       |           |
|      main |              0.1 |  10000 |   CRLF |        CRLF |  44.628 us |  1.00 |   20024 B |
|        pr |              0.1 |  10000 |   CRLF |        CRLF |  11.373 us |  0.25 |         - |
|           |                  |        |        |             |            |       |           |
|      main |              0.2 |  10000 |   CRLF |        CRLF |  70.777 us |  1.00 |   20024 B |
|        pr |              0.2 |  10000 |   CRLF |        CRLF |  18.566 us |  0.26 |         - |
|           |                  |        |        |             |            |       |           |
|      main |                1 |  10000 |   CRLF |        CRLF | 148.458 us |  1.00 |   20024 B |
|        pr |                1 |  10000 |   CRLF |        CRLF |  54.425 us |  0.37 |         - |

</details>

<details>
<summary>SpanLineEnumerator</summary>

|         Method | Toolchain | NewLineFrequency | Length |       Mean |     Error | Ratio |
|--------------- |---------- |----------------- |------- |-----------:|----------:|------:|
| EnumerateLines |      main |                0 |  10000 |   8.558 us | 0.0333 us |  1.00 |
| EnumerateLines |        pr |                0 |  10000 |   1.091 us | 0.0108 us |  0.13 |
|                |           |                  |        |            |           |       |
| EnumerateLines |      main |             0.05 |  10000 |  23.230 us | 0.0779 us |  1.00 |
| EnumerateLines |        pr |             0.05 |  10000 |   8.074 us | 0.0991 us |  0.35 |
|                |           |                  |        |            |           |       |
| EnumerateLines |      main |              0.1 |  10000 |  37.880 us | 0.1584 us |  1.00 |
| EnumerateLines |        pr |              0.1 |  10000 |  13.635 us | 0.0982 us |  0.36 |
|                |           |                  |        |            |           |       |
| EnumerateLines |      main |              0.2 |  10000 |  67.683 us | 0.3634 us |  1.00 |
| EnumerateLines |        pr |              0.2 |  10000 |  26.506 us | 0.1376 us |  0.39 |
|                |           |                  |        |            |           |       |
| EnumerateLines |      main |                1 |  10000 | 241.015 us | 1.0229 us |  1.00 |
| EnumerateLines |        pr |                1 |  10000 | 125.364 us | 1.4975 us |  0.52 |

</details>