@MihaZupan 2023-02-10
Any thoughts on this @dotnet/area-system-runtime?